requirements.txt
aspose-imaging-python-net==24.5.0
dithering==0.1.27
glitch-this==1.0.2
GlitchArt==1.0.0
imageio==2.33.1
joblib==1.3.2
lazy_loader==0.3
llvmlite==0.42.0rc1
networkx==3.2.1
numba==0.59.0rc1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyxelate==2.1.1
scikit-image==0.22.0
scikit-learn==1.4.0rc1
SciPy==1.12.0rc1
threadpoolctl==3.2.0
tifffile==2023.12.9
